Move remoteversion-related API tests
